added and fix hooks
