+++ src/components/Togglegenre/Togglegenre.js
@@ -1,7 +1,29 @@
 import React from "react";
 import "./ToggleGenre.styles.css";
 
-function ToggleGenre(props) {
+function ToggleGenre({
+  handler,
+  titleHandler,
+  setSortedMovieByYear,
+  setSortedMovieByTitle,
+}) {
+  //   const handlerYear = () => {
+  //     props.setSortedMovieByYear(false);
+  //     props.setSortedMovieByTitle(true);
+  //   };
+  //   const handlerTitle = () => {
+  //     props.setSortedMovieByYear(true);
+  //     props.setSortedMovieByTitle(false);
+  //   };
+
+  // const handlerYear = () => {
+  //   setSortedMovieByYear(false);
+  //   setSortedMovieByTitle(true);
+  // };
+  // const handlerTitle = () => {
+  //   setSortedMovieByYear(true);
+  //   setSortedMovieByTitle(false);
+  // };
   return (
     <div className="flex-row">
       <nav className="Navigator">
@@ -23,10 +45,10 @@ function ToggleGenre(props) {
       </nav>
       <div className="Sorting">
         <p className="Sorting-items">Sort by</p>
-        <p className="Sorting-items button" onClick={props.handler}>
+        <p className="Sorting-items button" onClick={handler}>
           Release date
         </p>
-        <p className="Sorting-items button" onClick={props.titleHandler}>
+        <p className="Sorting-items button" onClick={titleHandler}>
           Title
         </p>
       </div>
